refactor(gir): Move IntrospectedClassCallback to class-callback.ts and update imports
- Moved the `IntrospectedClassCallback` class to a new `class-callback.ts` file for better organization.
- Updated import statements across multiple files to reference the new location of `IntrospectedClassCallback`.
- Changed several imports to type-only imports to enhance clarity and maintainability.
