+++ packages/lib/src/generators/generator.ts
@@ -12,7 +12,8 @@ import type { IntrospectedFunctionParameter } from "../gir/parameter.ts";
 import type { IntrospectedConstructor } from "../gir/constructor.ts";
 import type { IntrospectedDirectAllocationConstructor } from "../gir/direct-allocation-constructor.ts";
 import type { IntrospectedFunction } from "../gir/function.ts";
-import type { IntrospectedCallback, IntrospectedClassCallback } from "../gir/callback.ts";
+import type { IntrospectedCallback } from "../gir/callback.ts";
+import type { IntrospectedClassCallback } from "../gir/class-callback.ts";
 import type { IntrospectedClassFunction } from "../gir/class-function.ts";
 import type { IntrospectedStaticClassFunction } from "../gir/static-function.ts";
 import type { IntrospectedVirtualClassFunction } from "../gir/virtual-function.ts";
