+++ packages/lib/src/generics/visitor.ts
@@ -6,18 +6,18 @@ import {
     GenerifiedTypeIdentifier,
     ThisType,
 } from "../gir.ts";
-import { IntrospectedEnum } from "../gir/enum.ts";
+import type { IntrospectedEnum } from "../gir/enum.ts";
 import { IntrospectedClass } from "../gir/introspected-class.ts";
 import { IntrospectedBaseClass } from "../gir/introspected-base-class.ts";
-import { IntrospectedInterface } from "../gir/interface.ts";
-import { IntrospectedFunctionParameter } from "../gir/parameter.ts";
+import type { IntrospectedInterface } from "../gir/interface.ts";
+import type { IntrospectedFunctionParameter } from "../gir/parameter.ts";
 import { IntrospectedFunction } from "../gir/function.ts";
 import { IntrospectedClassFunction } from "../gir/class-function.ts";
 import { IntrospectedStaticClassFunction } from "../gir/static-function.ts";
-import { IntrospectedVirtualClassFunction } from "../gir/virtual-function.ts";
-import { IntrospectedCallback } from "../gir/callback.ts";
+import type { IntrospectedVirtualClassFunction } from "../gir/virtual-function.ts";
+import type { IntrospectedCallback } from "../gir/callback.ts";
 import { GenericNameGenerator } from "../gir/generics.ts";
-import { NSRegistry } from "../gir/registry.ts";
+import type { NSRegistry } from "../gir/registry.ts";
 import { resolveTypeIdentifier } from "../utils/type-resolution.ts";
 import { GirVisitor } from "../visitor.ts";
