+++ packages/lib/src/gir/interface.ts
@@ -18,7 +18,7 @@ import {
     IntrospectedConstructor,
 } from "./constructor.ts";
 
-import { IntrospectedClassCallback } from "./callback.ts";
+import { IntrospectedClassCallback } from "./class-callback.ts";
 import { IntrospectedProperty, IntrospectedField } from "./property.ts";
 import { IntrospectedNamespace } from "./namespace.ts";
 import { sanitizeIdentifierName } from "../utils/naming.ts";
