+++ packages/lib/src/gir/signal.ts
@@ -14,7 +14,7 @@ import { isIntrospectable } from "../utils/girs.ts";
 import { IntrospectedClass } from "./introspected-class.ts";
 import { IntrospectedClassFunction } from "./class-function.ts";
 import { IntrospectedFunctionParameter } from "./parameter.ts";
-import { IntrospectedClassCallback } from "./callback.ts";
+import { IntrospectedClassCallback } from "./class-callback.ts";
 import type { GirSignalElement, GirCallableParamElement } from "@gi.ts/parser";
 import { GirDirection } from "@gi.ts/parser";
 import { getType, parseDoc, parseMetadata } from "../utils/gir-parsing.ts";
