+++ packages/lib/src/gir/virtual-function.ts
@@ -1,6 +1,6 @@
 import {
   UnknownType,
-  TypeExpression,
+  type TypeExpression,
 } from "../gir.ts";
 
 import type { IntrospectedOptions } from "../types/index.ts";
@@ -10,14 +10,14 @@ import type {
 } from "../index.ts";
 
 
-import { IntrospectedEnum } from "./enum.ts";
-import { FormatGenerator } from "../generators/generator.ts";
-import { GirVisitor } from "../visitor.ts";
-import { IntrospectedBaseClass } from "./introspected-base-class.ts";
+import type { IntrospectedEnum } from "./enum.ts";
+import type { FormatGenerator } from "../generators/generator.ts";
+import type { GirVisitor } from "../visitor.ts";
+import type { IntrospectedBaseClass } from "./introspected-base-class.ts";
 
 import type { OptionsLoad } from "../types/index.ts";
 
-import { IntrospectedFunctionParameter } from "./parameter.ts";
+import type { IntrospectedFunctionParameter } from "./parameter.ts";
 import { IntrospectedClassFunction } from "./class-function.ts";
 import { IntrospectedFunction } from "./function.ts";
