+++ packages/lib/src/visitor.ts
@@ -19,7 +19,8 @@ import type { IntrospectedFunction } from './gir/function.ts'
 import type { IntrospectedConstructor } from './gir/constructor.ts'
 import type { IntrospectedVirtualClassFunction } from './gir/virtual-function.ts'
 import type { IntrospectedDirectAllocationConstructor } from './gir/direct-allocation-constructor.ts'
-import type { IntrospectedCallback, IntrospectedClassCallback } from './gir/callback.ts'
+import type { IntrospectedCallback } from './gir/callback.ts'
+import type { IntrospectedClassCallback } from './gir/class-callback.ts'
 import type { IntrospectedNamespace } from './gir/namespace.ts'
 import type { IntrospectedProperty, IntrospectedField } from './gir/property.ts'
 import type { IntrospectedSignal, IntrospectedSignalType } from './gir/signal.ts'
